+++ package.json
@@ -151,9 +151,9 @@
     "prop-types": "^15.7.2",
     "react": "^17.0.1",
     "react-banner": "^1.0.0-rc.0",
-    "react-document-title": "^2.0.3",
     "react-dom": "^17.0.1",
     "react-g-analytics": "0.4.2",
+    "react-helmet-async": "^1.0.9",
     "react-router-dom": "^5.2.0",
     "react-spring": "next",
     "react-tiny-popover": "5",
