+++ src/index.jsx
@@ -1,13 +1,13 @@
 // Import External Dependencies
 import ReactDOM from 'react-dom';
-import { BrowserRouter, Route } from 'react-router-dom';
+import { BrowserRouter } from 'react-router-dom';
 import { BrowserRouter as AnalyticsRouter } from 'react-g-analytics';
 
-// Import Components
-import Site from './components/Site/Site';
+import App from './App.jsx';
 
 // Import helpers
 import isClient from './utilities/is-client';
+import { HelmetProvider } from 'react-helmet-async';
 
 const isProduction = process.env.NODE_ENV === 'production';
 
@@ -18,12 +18,9 @@ const render = isProduction ? ReactDOM.hydrate : ReactDOM.render;
 if (isClient) {
   render(
     <Router id="UA-46921629-2">
-      <Route
-        path="/"
-        render={(props) => (
-          <Site {...props} import={(path) => import(`./content/${path}`)} />
-        )}
-      />
+      <HelmetProvider>
+        <App />
+      </HelmetProvider>
     </Router>,
     document.getElementById('root')
   );
