+++ src/utilities/content-utils.js
@@ -88,3 +88,16 @@ export const getPageTitle = (tree, path) => {
 
   return title;
 };
+
+export const getPageDescription = (tree, path) => {
+  let page = findInContent(tree, (item) => item.url === path);
+  if (!page) {
+    return '';
+  } else if (path === '/') {
+    return '';
+  } else if (path.includes('/printable')) {
+    return '';
+  } else {
+    return page.description || '';
+  }
+};
