+++ webpack.ssg.js
@@ -11,7 +11,6 @@ const contentTree = require('./src/_content.json');
 
 // Load Common Configuration
 const common = require('./webpack.common.js');
-const PrecacheSsgManifestPlugin = require('./src/PrecacheSsgManifestPlugin');
 
 // content tree to path array
 const paths = [...flattenContentTree(contentTree), '/vote', '/app-shell'];
@@ -133,6 +132,5 @@ module.exports = (env) =>
           },
         ],
       }),
-      new PrecacheSsgManifestPlugin(),
     ],
   });
